+++ app/hero.service.js
@@ -12,6 +12,7 @@ var core_1 = require('@angular/core');
 var HeroService = (function () {
     function HeroService() {
     }
+    HeroService.prototype.getHeroes = function () { }; // stub
     HeroService = __decorate([
         core_1.Injectable(), 
         __metadata('design:paramtypes', [])
