+++ app/hero.service.ts
@@ -2,4 +2,5 @@ import { Injectable } from '@angular/core';
 
 @Injectable()
 export class HeroService {
+  getHeroes(): void {} // stub
 }
